Update README.md
You'll be working as a BI Developer for Northwind Traders, a global import and export company that
specializes in supplying high-quality gourmet food products to restaurants, cafes, and specialty food
retailers around the world.
As part of your role, you've been tasked with building a top-level KPI dashboard for the executive
team. Its purpose should be to allow them to quickly understand the company's performance in key
areas, including:
• Sales trends
• Product performance
• Key customers
• Shipping costs
The dashboard should be built to evolve and accommodate new data over time, but you've been
encouraged by your manager to have insights & recommendations ready to share with the VPs. This
is your chance to impress!
About The Data Set
Sales & order data for Northwind Traders, a fictitious gourmet food supplier, including information on
customers, products, orders, shippers, and employees.